% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07ObjetosS4.R
\name{addCol,DataFrame-method}
\alias{addCol,DataFrame-method}
\title{A?adir columna al dataFrame}
\usage{
\S4method{addCol}{DataFrame}(x, column, name = NULL)
}
\arguments{
\item{x}{El dataFrame}

\item{column}{La nueva columna que se quiere a?adir}

\item{name}{El nombre de la columna}
}
\value{
Un DataFrame con la columna a?adida
}
\description{
Esta funcion a?ade una columna a un objeto de tipo \code{\linkS4class{dataFrame}}
}
\examples{
DF <- dataFrame(list(a = c(1,2,3,3,5,7,8),
b = c("x", "y", "z", "x", "x", "y", "x"),
c = c(4,5,6,6,1,2,3),
d = c(7,8,9,7,6,5,3),
e = c(TRUE, TRUE, FALSE, FALSE, TRUE, FALSE, TRUE)))
DF = addCol(DF, rep(TRUE, 7))
}
