% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07ObjetosS4.R
\name{addRow,DataFrame-method}
\alias{addRow,DataFrame-method}
\title{A?adir fila al dataFrame}
\usage{
\S4method{addRow}{DataFrame}(x, line, name = NULL)
}
\arguments{
\item{x}{El dataFrame}

\item{line}{La nueva linea que se quiere a?adir}

\item{name}{El nombre de la fila}
}
\value{
Un DataFrame con la fila a?adida
}
\description{
Esta funcion a?ade una fila a un objeto de tipo \code{\linkS4class{dataFrame}}
}
\examples{
DF <- dataFrame(list(a = c(1,2,3,3,5,7,8),
b = c("x", "y", "z", "x", "x", "y", "x"),
c = c(4,5,6,6,1,2,3),
d = c(7,8,9,7,6,5,3),
e = c(TRUE, TRUE, FALSE, FALSE, TRUE, FALSE, TRUE)))
DF = addRow(DF, list(9, "x", 3, 3, FALSE))
}
