% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02Metricas.R
\name{calculoAUC}
\alias{calculoAUC}
\title{Calcular la area AUC}
\usage{
calculoAUC(df)
}
\arguments{
\item{df}{Un dataFrame con una columna numerica y una boolear}
}
\value{
El area AUC
}
\description{
Dado un dataFrame con una columna numerica y una boolear calcula su correspondiente area AUC
}
\examples{
df <- data.frame(x = sample(1:100, 20),
y = sample(c(TRUE, FALSE), 20, replace = TRUE))
calculoAUC(df)
}
