% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02Metricas.R
\name{calculoMetricas}
\alias{calculoMetricas}
\title{Calcular la varianza, AUC y entropia}
\usage{
calculoMetricas(df)
}
\arguments{
\item{df}{Un dataFrame con columnas categoricas y numericas. Las numericas deben estar seguidas por una logica}
}
\value{
Un vector con las varianzas, AUC y entropias
}
\description{
Dado un dataFrame con variables continuas y discretas, calcula la entropia de las discretas y AUC y varianza de las continuas
}
\examples{
df <- data.frame(x = sample(1:100, 20),
y = sample(c(TRUE, FALSE), 20, replace = TRUE))
df <- data.frame( a = as.factor(sample(c("a", "b", "c"), 15, replace = TRUE)),
b = sample(1:100, 15),
c = sample(c(TRUE, FALSE), 15, replace = TRUE),
d = as.factor(sample(c("a", "b", "c"), 15, replace = TRUE)),
e = as.factor(sample(c("a", "b", "c"), 15, replace = TRUE)),
f = sample(1:100, 15),
g = sample(c(TRUE, FALSE), 15, replace = TRUE))
calculoMetricas(df)
}
