% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02Metricas.R
\name{calculoVarianza}
\alias{calculoVarianza}
\title{Calcular la varianza}
\usage{
calculoVarianza(x)
}
\arguments{
\item{x}{Un vector numerico}
}
\value{
Los puntos de la curva ROC
}
\description{
Dado un vector numerico, calcula su varianza
}
\examples{
n <- 20
x <- round(runif(n)*100)
calculoVarianza(x)
}
