% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05Correlacion.R
\name{correlacion}
\alias{correlacion}
\title{Calcular la correlacion}
\usage{
correlacion(x, y)
}
\arguments{
\item{x}{El primer vector}

\item{y}{El segundo vector}
}
\value{
La correlacion o la informacion mutua
}
\description{
Dado dos vectores calcula la correlacion si los dos vectores son numericos, y la informacion mutua si uno de los dos vectores es categorico
}
\examples{
n <- 100
x <- sample(1:200, n, replace = TRUE)
y <- sample(1:200, n, replace = TRUE)
correlacion(x,y)
}
