% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05Correlacion.R
\name{correlacionDF}
\alias{correlacionDF}
\title{Calcular la correlacion de un dataFrame}
\usage{
correlacionDF(df)
}
\arguments{
\item{df}{El dataFrame}
}
\value{
Las matriz de correlaciones
}
\description{
Dado un dataFrame calcula la correlacion por pares entre variables
}
\examples{
x <- sample(1:500, 100, replace = TRUE)
df <- data.frame(a = sample(1:500, 100, replace = TRUE),
b = 1:100,
c = sample(1:500, 100, replace = TRUE),
d = seq(1, 1000, 10))
correlacionDF(df)
}
