% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05Correlacion.R
\name{correlacionUnaVariable}
\alias{correlacionUnaVariable}
\title{Calcular la correlacion de una variable}
\usage{
correlacionUnaVariable(x, df)
}
\arguments{
\item{x}{El vector}

\item{df}{El dataFrame}
}
\value{
Las correlaciones que mantiene x con todas las columnas del dataFrame
}
\description{
Dado un vector x y un dataFrame, calcula las correlaciones que mantiene x con todas las columnas del dataFrame
}
