% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02Metricas.R
\name{curvaROC}
\alias{curvaROC}
\title{Calcular la curva ROC}
\usage{
curvaROC(df)
}
\arguments{
\item{df}{Un dataFrame con una columna numerica y una boolear}
}
\value{
Los puntos de la curva ROC
}
\description{
Dado un dataFrame con una columna numerica y una boolear calcula los puntos de su curva ROC
}
