% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07ObjetosS4.R
\docType{class}
\name{DataFrame-class}
\alias{DataFrame-class}
\title{Una clase S4 para representar dataFrames}
\description{
Una clase S4 para representar dataFrames
}
\section{Slots}{

\describe{
\item{\code{colNames}}{Los nombres de las columnas}

\item{\code{rowNames}}{Los nombres de las filas}

\item{\code{colNum}}{Numero de columnas}

\item{\code{rowNum}}{Numero de filas}

\item{\code{data}}{La lista que guarda todos los datos}
}}

