% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07ObjetosS4.R
\name{dataFrame}
\alias{dataFrame}
\title{Constructor de la clase DataFrame}
\usage{
dataFrame(data, colNames = NULL, rowNames = NULL)
}
\arguments{
\item{data}{Lista que guarda toda la informacion}

\item{colNames}{Los nombres de las columnas}

\item{rowNames}{Los nombres de las filas}
}
\value{
Un objeto de la clase \code{\linkS4class{DataFrame}}
}
\description{
Esta funcion crea una instancia de \code{\linkS4class{dataFrame}}
}
