% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01Discretizacion.R
\name{discretizarColumnas}
\alias{discretizarColumnas}
\title{Discretizar dataFrame mediante Jenks Natural Breaks}
\usage{
discretizarColumnas(df, num.bins, discretizacion = "TODAS")
}
\arguments{
\item{df}{El dataFrame a discretizar}

\item{num.bins}{Numero de intervalos que se desean crear}

\item{discretizacion}{El metodo de discretizacion}
}
\value{
Un dataFrame con columnas categoricas resultado de aplicar la discretizacion aplicada a ellas
}
\description{
Dado un dataFrame con columnas numericas y un numero de intervalos esta funciun discretiza todas las columnas utilizando la estrategia indicada
}
\examples{
n <- 50
df <- data.frame(x = round(runif(n)*100),
y = round(runif(n)*100),
z = round(runif(n)*100))
num.bins <- 4
discretizarColumnas(df,num.bins,"EW")
}
