% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01Discretizacion.R
\name{discretize}
\alias{discretize}
\title{Discretizar atributo a corde a unos puntos de corte}
\usage{
discretize(x, cut.points)
}
\arguments{
\item{x}{El vector numerico a discretizar}

\item{cut.points}{El vector de puntos de corte}
}
\value{
Un vector de valores categoricos resultado de aplicar la discretizacion
}
\description{
Dado un vector de tipo numerico y un vector de puntos de corte, discretiza el vector numerico utilizando los puntos de corte
}
