% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01Discretizacion.R
\name{discretizeEW}
\alias{discretizeEW}
\title{Discretizar atributo mediante Equal Width}
\usage{
discretizeEW(x, num.bins)
}
\arguments{
\item{x}{El vector numerico a discretizar}

\item{num.bins}{Numero de intervalos que se desean crear}
}
\value{
Un vector de valores categoricos resultado de aplicar el Equal Width y un vector numerico que guarda los puntos de corte
}
\description{
Dado un vector de tipo numerico y un numero de intervalos esta funciun discretiza el vector utilizando la estrategia Equal Width
}
\examples{
x <- sample(1:100, 50)
num.bins <- 5
discretizeEW(x, num.bins)
}
