% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01Discretizacion.R
\name{discretizeJNB}
\alias{discretizeJNB}
\title{Discretizar atributo mediante Jenks Natural Breaks}
\usage{
discretizeJNB(x, num.bins)
}
\arguments{
\item{x}{El vector numerico a discretizar}

\item{num.bins}{Numero de intervalos que se desean crear}
}
\value{
Un vector de valores categoricos resultado de aplicar el Jenks Natural Breaks y un vector numerico que guarda los puntos de corte
}
\description{
Dado un vector de tipo numerico y un numero de intervalos esta funciun discretiza el vector utilizando la estrategia Jenks Natural Breaks
}
\examples{
x <- sample(1:100, 50)
num.bins <- 5
discretizeJNB(x, num.bins)
}
