% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02Metricas.R
\name{entropy}
\alias{entropy}
\title{Calcular la entropia de un vector}
\usage{
entropy(x.discretized)
}
\arguments{
\item{x.discretized}{El vector de entrada}
}
\value{
La entropia del vector
}
\description{
Dado un vector de tipo numerico o categorico calcula la entropia
}
\examples{
x.discretized <- factor(c("a", "c", "c", "a", "c"), levels = c("a", "b", "c"))
entropy(x.discretized)
}
