% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03NormalizacionYEstandarizacion.R
\name{estandarizar}
\alias{estandarizar}
\title{Estandarizar un vector}
\usage{
estandarizar(x)
}
\arguments{
\item{x}{El vector numerico a estandarizar}
}
\value{
El vector estandarizado
}
\description{
Dado un vector de tipo numerico estandariza dicho vector
}
\examples{
x <- sample(1:100, 50)
xest <- estandarizar(x)
}
