% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03NormalizacionYEstandarizacion.R
\name{estandarizarDF}
\alias{estandarizarDF}
\title{Estandarizar un dataFrame}
\usage{
estandarizarDF(df)
}
\arguments{
\item{df}{El dataFrame a estandarizar}
}
\value{
El dataFrame estandarizado
}
\description{
Dado un dataFrame con columnas numericas estandariza dicho dataFrame
}
\examples{
df <- data.frame( x = sample(1:100, 50),
y = sample(1:100, 50),
z = sample(1:100, 50))
dfest <- estandarizarDF(df)
}
