% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04Filtrado.R
\name{filtrar}
\alias{filtrar}
\title{Filtrar un dataFrame}
\usage{
filtrar(
  df,
  varianza = FALSE,
  AUC = FALSE,
  ent = FALSE,
  varUmbr = 5,
  AUCUmbr = 0.5,
  entUmbr = 0.5,
  varOP = "EQ",
  AUCOP = "EQ",
  entOP = "EQ"
)
}
\arguments{
\item{df}{El dataFrame a filtrar}

\item{varianza}{Un valor logico que indica si se quiere filtrar por la varianza}

\item{AUC}{Un valor logico que indica si se quiere filtrar por el AUC}

\item{ent}{Un valor logico que indica si se quiere filtrar por la entropia}

\item{varUmbr}{El umbral de la varianza}

\item{AUCUmbr}{El umbral del AUC}

\item{entUmbr}{El umbral de la entropia}

\item{varOP}{El operador que se utilizara al comparar la varianza}

\item{AUCOP}{El operador que se utilizara al comparar el AUC}

\item{entOP}{El operador que se utilizara al comparar la entropia}
}
\value{
El dataFrame filtrado
}
\description{
Dado un dataFrame filtra las variables segun la entropia, el AUC o la varianza
}
\examples{
n <- 20
df <- data.frame(a.x = sample(1:100, n, replace = TRUE),
a.y = sample(c(TRUE, FALSE), n, replace = TRUE),
b = sample(c("a", "b", "c", "d"), n, replace = TRUE),
c.x = sample(1:50, n, replace = TRUE),
c.y = sample(c(TRUE, FALSE), n, replace = TRUE),
d = sample(c("a", "b", "c"), n, replace = TRUE),
e = sample(c("a", "b"), n, replace = TRUE),
f = sample(50:100, n, replace = TRUE))
filtrar(df, ent = TRUE, entUmbr = 1.6, entOP = "LO")
}
