% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07ObjetosS4.R
\name{getCols,DataFrame-method}
\alias{getCols,DataFrame-method}
\title{Conseguir columnas de dataFrame}
\usage{
\S4method{getCols}{DataFrame}(x, names)
}
\arguments{
\item{x}{El dataFrame}

\item{names}{Los nombres de las columnas}
}
\value{
Un DataFrame que contiene solo las columnas indicadas
}
\description{
Esta funcion consigue las columnas indicadas de un objeto de tipo \code{\linkS4class{dataFrame}}
}
\examples{
DF <- dataFrame(list(a = c(1,2,3,3,5,7,8),
b = c("x", "y", "z", "x", "x", "y", "x"),
c = c(4,5,6,6,1,2,3),
d = c(7,8,9,7,6,5,3),
e = c(TRUE, TRUE, FALSE, FALSE, TRUE, FALSE, TRUE)))
DFaux <- getCols(DF, c("a", "c"))
}
