% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07ObjetosS4.R
\name{getRows,DataFrame-method}
\alias{getRows,DataFrame-method}
\title{Conseguir filas del dataFrame}
\usage{
\S4method{getRows}{DataFrame}(x, names)
}
\arguments{
\item{x}{El dataFrame}

\item{names}{Los nombres de las filas}
}
\value{
Un DataFrame que contiene solo las filas indicadas
}
\description{
Esta funcion consigue las filas indicadas de un objeto de tipo \code{\linkS4class{dataFrame}}
}
\examples{
DF <- dataFrame(list(a = c(1,2,3,3,5,7,8),
b = c("x", "y", "z", "x", "x", "y", "x"),
c = c(4,5,6,6,1,2,3),
d = c(7,8,9,7,6,5,3),
e = c(TRUE, TRUE, FALSE, FALSE, TRUE, FALSE, TRUE)))
DFaux <- getRows(DF, c("y3", "y1"))
}
