% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03NormalizacionYEstandarizacion.R
\name{normalizar}
\alias{normalizar}
\title{Normalizar un vector}
\usage{
normalizar(x)
}
\arguments{
\item{x}{El vector numerico a normalizar}
}
\value{
El vector normalizado
}
\description{
Dado un vector de tipo numerico normaliza dicho vector
}
\examples{
x <- sample(1:100, 50)
xnorm <- normalizar(x)
}
