% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03NormalizacionYEstandarizacion.R
\name{normalizarDF}
\alias{normalizarDF}
\title{Normalizar un dataFrame}
\usage{
normalizarDF(df)
}
\arguments{
\item{df}{El dataFrame a normalizar}
}
\value{
El dataFrame normalizado
}
\description{
Dado un dataFrame con columnas numericas normaliza dicho dataFrame
}
\examples{
df <- data.frame( x = sample(1:100, 50),
y = sample(1:100, 50),
z = sample(1:100, 50))
dfnorm <- normalizarDF(df)
}
