% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06Visualizacion.R
\name{plotBarrasDiscretizadas}
\alias{plotBarrasDiscretizadas}
\title{Visualizar barras discretizadas}
\usage{
plotBarrasDiscretizadas(x, num.bins, metodo = "EW")
}
\arguments{
\item{x}{El vector numerico}

\item{num.bins}{El numero de intervalos}

\item{metodo}{El metodo de discretizacion (EW, EB o JNB)}
}
\description{
Dado un vector numerico, un numero de intervalos y un metodo de discretizacion, discretiza el vector y muestra la frecuancia de todos los intervalos
}
\examples{
x<-sample(1:1000, 100)
num.bins <- 5
plotBarrasDiscretizadas(x, num.bins, "EF")
}
