% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06Visualizacion.R
\name{plotBoxPlots}
\alias{plotBoxPlots}
\title{Visualizar boxPlots}
\usage{
plotBoxPlots(df)
}
\arguments{
\item{df}{El dataFrame}
}
\description{
Dado un dataFrame con variables numericas visualiza todos los boxplots de las variables
}
\examples{
df <- data.frame(a = sample(1:500, 100, replace = TRUE),
b = sample(1:200, 100, replace = TRUE),
c = sample(100:800, 100, replace = TRUE),
d = sample(1:1000, 100, replace = TRUE))
plotBoxPlots(df)
}
