% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06Visualizacion.R
\name{plotCorrelaciones}
\alias{plotCorrelaciones}
\title{Visualizar correlaciones}
\usage{
plotCorrelaciones(df)
}
\arguments{
\item{df}{El dataFrame}
}
\description{
Dado un dataFrame calcula la correlacion por pares entre variables
}
\examples{
library(ggplot2)
library(reshape2)
df <- data.frame(a = sample(1:500, 100, replace = TRUE),
b = 1:100,
c = sample(1:500, 100, replace = TRUE),
d = seq(1, 1000, 10))
plotCorrelaciones(df)
}
