% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06Visualizacion.R
\name{plotROC}
\alias{plotROC}
\title{Visualizar curva ROC}
\usage{
plotROC(df)
}
\arguments{
\item{df}{El dataFrame}
}
\description{
Dado un dataFrame con una columna numerica y una boolear visualiza su curva ROC
}
\examples{
df <- data.frame(a = sample(1:500, 100, replace = TRUE),
b = sample(c(FALSE, TRUE), 100, replace = TRUE))
plotROC(df)
}
