% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07ObjetosS4.R
\name{print,DataFrame-method}
\alias{print,DataFrame-method}
\title{Imprimir dataFrame}
\usage{
\S4method{print}{DataFrame}(x)
}
\arguments{
\item{x}{El dataFrame}
}
\description{
Esta función imprime un objeto de tipo \code{\linkS4class{dataFrame}}
}
\examples{
DF <- dataFrame(list(a = c(1,2,3,3,5,7,8),
b = c("x", "y", "z", "x", "x", "y", "x"),
c = c(4,5,6,6,1,2,3),
d = c(7,8,9,7,6,5,3),
e = c(TRUE, TRUE, FALSE, FALSE, TRUE, FALSE, TRUE)))
}
