% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08LecturaEscrituraData.R
\name{readDataFrame}
\alias{readDataFrame}
\title{Leer un dataFrame de un fichero csv}
\usage{
readDataFrame(file.path, header = FALSE, rownames = FALSE, sep = "\\t")
}
\arguments{
\item{file.path}{Directorio del fichero}

\item{header}{Indicativo de si el fichero guarda los nombres de las columnas}

\item{rownames}{Indicativo de si el fichero guarda los nombres de las filas}

\item{sep}{Separador que se ha utilizado en el fichero csv}
}
\value{
El dataFrame con toda la informacion
}
\description{
Dado un directorio de un fichero lee la informacion y crea un dataFrame
}
\examples{
n <- 10
df <- data.frame( a = round(runif(n) * 100 ,2),
b = as.character(letters[1:n]),
c = as.factor(sample(c("a", "b", "c"), n, replace = TRUE)),
d = sample(1:100, n),
e = round(runif(n) * 100 ,2),
f = as.factor(sample(c("a", "b", "c"), n, replace = TRUE)))
writeDataFrame(df, "tabla.csv", header = FALSE, rownames = FALSE, sep = "#")
df <- readDataFrame("tabla.csv", header = FALSE, rownames = FALSE, sep = "#")
}
