% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02Metricas.R
\name{TPRyFPR}
\alias{TPRyFPR}
\title{Calcular TPR y FPR}
\usage{
TPRyFPR(valor.corte, df)
}
\arguments{
\item{valor.corte}{El valor de corte}

\item{df}{Un dataFrame con una columna numerica y una boolear}
}
\value{
La entropia del vector
}
\description{
Dado un dataFrame con una columna numerica y una boolear y un valor de corte, calula el TPR y FPR
}
