% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08LecturaEscrituraData.R
\name{writeDataFrame}
\alias{writeDataFrame}
\title{Escribir un dataFrame en un fichero csv}
\usage{
writeDataFrame(
  dataFrame,
  file.path,
  header = FALSE,
  rownames = FALSE,
  sep = "\\t"
)
}
\arguments{
\item{dataFrame}{El dataFrame}

\item{file.path}{Directorio del fichero}

\item{header}{Indicativo de si se quieren guardar los nombres de las columnas}

\item{rownames}{Indicativo de si se quieren guardar los nombres de las filas}

\item{sep}{Separador que se quiere utilizar en el fichero csv}
}
\description{
Dado un directorio de un fichero y un dataFrame, escribe el dataFrame en el fichero
}
\examples{
n <- 10
df <- data.frame( a = round(runif(n) * 100 ,2),
b = as.character(letters[1:n]),
c = as.factor(sample(c("a", "b", "c"), n, replace = TRUE)),
d = sample(1:100, n),
e = round(runif(n) * 100 ,2),
f = as.factor(sample(c("a", "b", "c"), n, replace = TRUE)))
writeDataFrame(df, "tabla.csv", header = FALSE, rownames = FALSE, sep = "#")
}
